+++ src/libraries/System.Private.CoreLib/src/System/Diagnostics/Tracing/IncrementingEventCounter.cs
@@ -12,9 +12,6 @@ namespace System.Diagnostics.Tracing
     /// It does not calculate statistics like mean, standard deviation, etc. because it only accumulates
     /// the counter value.
     /// </summary>
-#if !ES_BUILD_STANDALONE
-    [UnsupportedOSPlatform("browser")]
-#endif
     public partial class IncrementingEventCounter : DiagnosticCounter
     {
         /// <summary>
